+++ packages/react/runtime/src/index.ts
@@ -413,7 +413,14 @@ Object.defineProperties(Signal.prototype, {
 	props: {
 		configurable: true,
 		get() {
-			return { data: this };
+			const s: Signal = this;
+			return {
+				data: {
+					get value() {
+						return s.value;
+					},
+				},
+			};
 		},
 	},
 	ref: { configurable: true, value: null },
@@ -465,8 +472,10 @@ declare global {
 }
 
 /** See comment in packages/core/src/index.ts on the same interface for an explanation */
-interface InternalModelConstructor<TModel, TArgs extends any[]>
-	extends ModelConstructor<TModel, TArgs> {
+interface InternalModelConstructor<
+	TModel,
+	TArgs extends any[],
+> extends ModelConstructor<TModel, TArgs> {
 	(...args: TArgs): Model<TModel>;
 }